[refactor] 디렉터리 구조 수정
가독성 향상
