+++ services/Pripen/src/main/java/com/example/backend/controller/UserController.java
@@ -4,21 +4,22 @@ package com.example.backend.controller;
 import com.example.backend.dto.CompanyDTO;
 import com.example.backend.dto.ResponseMessage;
 import com.example.backend.dto.UserDto;
-import com.example.backend.model.Company;
-import com.example.backend.model.User;
+import com.example.backend.model.Authentication.Company;
+import com.example.backend.model.Authentication.User;
 import com.example.backend.model.redis.RefreshToken;
 import com.example.backend.repository.redis.RefreshTokenRepository;
-import com.example.backend.service.*;
+import com.example.backend.service.Authentication.Login.*;
+import com.example.backend.service.Authentication.Signup.CompanyService;
+import com.example.backend.service.Authentication.Signup.EmailPostService;
+import com.example.backend.service.Authentication.Signup.EmailVerificationService;
 import jakarta.servlet.http.Cookie;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import org.modelmapper.ModelMapper;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.authentication.AuthenticationManager;
 import org.springframework.security.authentication.BadCredentialsException;
-import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
 import org.springframework.security.core.Authentication;
 import org.springframework.security.core.AuthenticationException;
 import org.springframework.security.core.context.SecurityContextHolder;
