+++ services/Pripen/src/main/java/com/example/backend/model/Analysis/AnalysisIssuesGuidelines.java
@@ -1,0 +1,4 @@
+package com.example.backend.model.Analysis;
+
+public class AnalysisIssuesGuidelines {
+}
