+++ services/Pripen/src/main/java/com/example/backend/model/Analysis/AnalysisIssuesReasons.java
@@ -1,0 +1,4 @@
+package com.example.backend.model.Analysis;
+
+public class AnalysisIssuesReasons {
+}
