+++ services/Pripen/src/main/java/com/example/backend/model/CustomUserDetails.java
@@ -1,5 +1,6 @@
 package com.example.backend.model;
 
+import com.example.backend.model.Authentication.User;
 import org.springframework.security.core.GrantedAuthority;
 import org.springframework.security.core.userdetails.UserDetails;
 import java.util.Collection;
