+++ services/Pripen/src/main/java/com/example/backend/repository/CompanyRepository.java
@@ -1,6 +1,6 @@
 package com.example.backend.repository;
 
-import com.example.backend.model.Company;
+import com.example.backend.model.Authentication.Company;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
