+++ services/Pripen/src/main/java/com/example/backend/repository/UserRepository.java
@@ -2,7 +2,7 @@ package com.example.backend.repository;
 
 
 
-import com.example.backend.model.User;
+import com.example.backend.model.Authentication.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
