+++ services/Pripen/src/main/java/com/example/backend/service/Authentication/Login/CookieService.java
@@ -1,4 +1,4 @@
-package com.example.backend.service;
+package com.example.backend.service.Authentication.Login;
 
 import jakarta.servlet.http.Cookie;
 import jakarta.servlet.http.HttpServletResponse;
