+++ services/Pripen/src/main/java/com/example/backend/service/Authentication/Login/JWTService.java
@@ -1,4 +1,4 @@
-package com.example.backend.service;
+package com.example.backend.service.Authentication.Login;
 
 import io.jsonwebtoken.*;
 import org.slf4j.Logger;
