+++ services/Pripen/src/main/java/com/example/backend/service/Authentication/Login/UserDetailServiceImpl.java
@@ -1,7 +1,7 @@
-package com.example.backend.service;
+package com.example.backend.service.Authentication.Login;
 
 import com.example.backend.model.CustomUserDetails;
-import com.example.backend.model.User;
+import com.example.backend.model.Authentication.User;
 import com.example.backend.repository.UserRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.core.userdetails.UserDetails;
